+++ x/evm/types/msg_test.go
@@ -120,7 +120,7 @@ func (suite *MsgsTestSuite) TestMsgEthereumTx_BuildTx() {
 		for _, tc := range testCases {
 			configurator := types.NewEVMConfigurator()
 			configurator.ResetTestChainConfig()
-			suite.Require().NoError(configurator.WithEVMCoinInfo(cfg.Denom, int64(cfg.Decimals)).Configure())
+			suite.Require().NoError(configurator.WithEVMCoinInfo(cfg.Denom, uint32(cfg.Decimals)).Configure())
 			if strings.Contains(tc.name, "nil data") {
 				tc.msg.Data = nil
 			}
